Fix Metal compatibility for Intel Mac AMD GPU (macOS 11.1 SDK)
Three fixes for building and running on Intel Macs with AMD Radeon GPUs
using the macOS 11.1 SDK:

- MTLResource.hpp: add fallback typedef for task_id_token_t, which was
  introduced in the macOS 13 SDK and is absent from older SDKs

- plume_apple.mm: fix cast of [NSScreen maximumFramesPerSecond] (macOS 12+
  API) through NSInteger to avoid pointer-to-int error with SDK 11.1,
  where the undeclared method defaults to returning id

- plume_metal.cpp: on Tier 2 devices without Metal 3 (e.g. Intel Mac AMD),
  call setArgumentBuffer on the encoder before setBuffer in
  MetalDescriptorSet::setDescriptor. macOS 11+ reports all GPUs as Tier 2,
  but Intel AMD GPUs do not support direct buffer addresses; without
  setArgumentBuffer the encoder has no backing buffer, causing a SIGSEGV
  inside AMDMTLBronzeDriver.

+++ plume_apple.mm
@@ -72,7 +72,7 @@ namespace plume {
 
             NSScreen *screen = [nsWindow screen];
             if (@available(macOS 12.0, *)) {
-                cachedRefreshRate.store((int)[screen maximumFramesPerSecond]);
+                cachedRefreshRate.store((int)(NSInteger)[screen maximumFramesPerSecond]);
             }
         } else {
             updateWindowAttributesInternal(true);
@@ -107,7 +107,7 @@ namespace plume {
             NSWindow *nsWindow = (__bridge NSWindow *)windowHandle;
             NSScreen *screen = [nsWindow screen];
             if (@available(macOS 12.0, *)) {
-                cachedRefreshRate.store((int)[screen maximumFramesPerSecond]);
+                cachedRefreshRate.store((int)(NSInteger)[screen maximumFramesPerSecond]);
             }
         };
 
@@ -149,7 +149,7 @@ namespace plume {
             NSScreen *screen = [nsWindow screen];
 
             if (@available(macOS 12.0, *)) {
-                int freshRate = (int)[screen maximumFramesPerSecond];
+                int freshRate = (int)(NSInteger)[screen maximumFramesPerSecond];
                 const_cast<CocoaWindow*>(this)->cachedRefreshRate.store(freshRate);
                 return freshRate;
             }
